Squashed commits

+++ scripts/clean.sh
@@ -12,5 +12,11 @@ rm -f app/shared/tsconfig.tsbuildinfo
 rm -f app/server/tsconfig.tsbuildinfo
 rm -f app/web/tsconfig.tsbuildinfo
 
-# This deletes build cache, don't run too often
-docker system prune --volumes -f
+if [[ ! " $@ " =~ " --no-docker " ]]; then
+  read -n 1 -p "Prune Docker? " -r
+  echo ""
+  if [[ $REPLY =~ ^[Yy]$ ]]; then
+    # This deletes build cache, don't run too often
+    docker system prune --volumes -f
+  fi
+fi
